KG-146. Support subgraph events in Debugger feature
