+++ agents/agents-core/src/jvmTest/kotlin/ai/koog/agents/core/system/feature/DebuggerTest.kt
@@ -6,22 +6,16 @@ import ai.koog.agents.core.dsl.builder.forwardTo
 import ai.koog.agents.core.dsl.builder.strategy
 import ai.koog.agents.core.dsl.extension.nodeExecuteTool
 import ai.koog.agents.core.dsl.extension.nodeLLMRequest
-import ai.koog.agents.core.dsl.extension.nodeLLMRequestStreamingAndSendResults
 import ai.koog.agents.core.dsl.extension.nodeLLMSendToolResult
 import ai.koog.agents.core.dsl.extension.onAssistantMessage
 import ai.koog.agents.core.dsl.extension.onToolCall
 import ai.koog.agents.core.environment.ReceivedToolResult
 import ai.koog.agents.core.feature.debugger.Debugger
-import ai.koog.agents.core.feature.model.AIAgentError
 import ai.koog.agents.core.feature.model.events.AgentCompletedEvent
 import ai.koog.agents.core.feature.model.events.AgentStartingEvent
 import ai.koog.agents.core.feature.model.events.GraphStrategyStartingEvent
 import ai.koog.agents.core.feature.model.events.LLMCallCompletedEvent
 import ai.koog.agents.core.feature.model.events.LLMCallStartingEvent
-import ai.koog.agents.core.feature.model.events.LLMStreamingCompletedEvent
-import ai.koog.agents.core.feature.model.events.LLMStreamingFailedEvent
-import ai.koog.agents.core.feature.model.events.LLMStreamingFrameReceivedEvent
-import ai.koog.agents.core.feature.model.events.LLMStreamingStartingEvent
 import ai.koog.agents.core.feature.model.events.NodeExecutionCompletedEvent
 import ai.koog.agents.core.feature.model.events.NodeExecutionStartingEvent
 import ai.koog.agents.core.feature.model.events.StrategyCompletedEvent
@@ -34,10 +28,15 @@ import ai.koog.agents.core.feature.remote.client.FeatureMessageRemoteClient
 import ai.koog.agents.core.feature.remote.client.config.DefaultClientConnectionConfig
 import ai.koog.agents.core.feature.remote.server.config.DefaultServerConnectionConfig
 import ai.koog.agents.core.feature.writer.FeatureMessageRemoteWriter
+import ai.koog.agents.core.system.feature.DebuggerTestAPI.HOST
+import ai.koog.agents.core.system.feature.DebuggerTestAPI.connectWithRetry
+import ai.koog.agents.core.system.feature.DebuggerTestAPI.mockLLModel
+import ai.koog.agents.core.system.feature.DebuggerTestAPI.runAgentConnectionWaitConfigThroughSystemVariablesTest
+import ai.koog.agents.core.system.feature.DebuggerTestAPI.runAgentPortConfigThroughSystemVariablesTest
+import ai.koog.agents.core.system.feature.DebuggerTestAPI.testBaseClient
 import ai.koog.agents.core.system.getEnvironmentVariableOrNull
 import ai.koog.agents.core.system.getVMOptionOrNull
 import ai.koog.agents.core.system.mock.ClientEventsCollector
-import ai.koog.agents.core.system.mock.MockLLMProvider
 import ai.koog.agents.core.system.mock.assistantMessage
 import ai.koog.agents.core.system.mock.createAgent
 import ai.koog.agents.core.system.mock.systemMessage
@@ -45,39 +44,24 @@ import ai.koog.agents.core.system.mock.testClock
 import ai.koog.agents.core.system.mock.toolCallMessage
 import ai.koog.agents.core.system.mock.toolResultMessage
 import ai.koog.agents.core.system.mock.userMessage
-import ai.koog.agents.core.tools.ToolDescriptor
 import ai.koog.agents.core.tools.ToolRegistry
 import ai.koog.agents.core.utils.SerializationUtils
 import ai.koog.agents.testing.network.NetUtil
 import ai.koog.agents.testing.network.NetUtil.findAvailablePort
 import ai.koog.agents.testing.tools.DummyTool
 import ai.koog.agents.testing.tools.getMockExecutor
-import ai.koog.prompt.dsl.ModerationResult
 import ai.koog.prompt.dsl.Prompt
-import ai.koog.prompt.executor.model.PromptExecutor
-import ai.koog.prompt.llm.LLModel
 import ai.koog.prompt.llm.toModelInfo
 import ai.koog.prompt.message.Message
 import ai.koog.prompt.message.RequestMetaInfo
-import ai.koog.prompt.streaming.StreamFrame
 import ai.koog.utils.io.use
-import io.ktor.client.HttpClient
-import io.ktor.client.plugins.HttpRequestRetry
 import io.ktor.http.URLProtocol
-import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CompletableDeferred
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.withContext
-import kotlinx.coroutines.withTimeout
 import kotlinx.coroutines.withTimeoutOrNull
-import kotlinx.io.IOException
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.parallel.Execution
@@ -87,7 +71,6 @@ import kotlin.reflect.typeOf
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
-import kotlin.test.assertFails
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
@@ -105,33 +88,7 @@ class DebuggerTest {
 
     companion object {
         private val defaultClientServerTimeout = 30.seconds
-        private const val HOST = "127.0.0.1"
     }
-
-    private suspend fun FeatureMessageRemoteClient.connectWithRetry(timeout: Duration) {
-        withTimeout(timeout) {
-            while (true) {
-                try {
-                    connect()
-                    return@withTimeout
-                } catch (exception: Exception) {
-                    if (exception is CancellationException) {
-                        throw exception
-                    }
-                    delay(100)
-                }
-            }
-        }
-    }
-
-    private val testBaseClient: HttpClient
-        get() = HttpClient {
-            install(HttpRequestRetry) {
-                retryOnExceptionIf(maxRetries = 10) { _, cause ->
-                    cause is IOException
-                }
-            }
-        }
 
     @AfterEach
     fun cleanup() {
@@ -164,15 +121,6 @@ class DebuggerTest {
         val toolRegistry = ToolRegistry {
             tool(dummyTool)
         }
-
-        // Model
-        val modelId = "test-llm-id"
-        val testModel = LLModel(
-            provider = MockLLMProvider(),
-            id = modelId,
-            capabilities = emptyList(),
-            contextLength = 1_000,
-        )
 
         // Prompt
         val promptId = "Test prompt id"
@@ -246,7 +194,7 @@ class DebuggerTest {
                 assistantPrompt = assistantPrompt,
                 toolRegistry = toolRegistry,
                 promptExecutor = mockExecutor,
-                model = testModel,
+                model = mockLLModel,
             ) {
                 @OptIn(ExperimentalAgentsApi::class)
                 install(Debugger) {
@@ -379,7 +327,7 @@ class DebuggerTest {
                         runId = clientEventsCollector.runId,
                         callId = callIds[0],
                         prompt = expectedLLMCallPrompt,
-                        model = testModel.toModelInfo(),
+                        model = mockLLModel.toModelInfo(),
                         tools = listOf(dummyTool.name),
                         timestamp = testClock.now().toEpochMilliseconds()
                     ),
@@ -387,7 +335,7 @@ class DebuggerTest {
                         runId = clientEventsCollector.runId,
                         callId = callIds[0],
                         prompt = expectedLLMCallPrompt,
-                        model = testModel.toModelInfo(),
+                        model = mockLLModel.toModelInfo(),
                         responses = listOf(toolCallMessage(dummyTool.name, content = """{"dummy":"$requestedDummyToolArgs"}""")),
                         timestamp = testClock.now().toEpochMilliseconds()
                     ),
@@ -452,7 +400,7 @@ class DebuggerTest {
                         runId = clientEventsCollector.runId,
                         callId = callIds[1],
                         prompt = expectedLLMCallWithToolsPrompt,
-                        model = testModel.toModelInfo(),
+                        model = mockLLModel.toModelInfo(),
                         tools = listOf(dummyTool.name),
                         timestamp = testClock.now().toEpochMilliseconds()
                     ),
@@ -460,7 +408,7 @@ class DebuggerTest {
                         runId = clientEventsCollector.runId,
                         callId = callIds[1],
                         prompt = expectedLLMCallWithToolsPrompt,
-                        model = testModel.toModelInfo(),
+                        model = mockLLModel.toModelInfo(),
                         responses = listOf(assistantMessage(mockResponse)),
                         timestamp = testClock.now().toEpochMilliseconds()
                     ),
@@ -532,368 +480,6 @@ class DebuggerTest {
         assertNotNull(isFinishedOrNull, "Client or server did not finish in time")
     }
 
-    @Test
-    fun `test feature message remote writer collect streaming success events on agent run`() = runBlocking {
-        // Agent Config
-        val agentId = "test-agent-id"
-
-        val userPrompt = "Call the dummy tool with argument: test"
-        val systemPrompt = "Test system prompt"
-        val assistantPrompt = "Test assistant prompt"
-        val promptId = "Test prompt id"
-
-        // Tools
-        val dummyTool = DummyTool()
-
-        val toolRegistry = ToolRegistry {
-            tool(dummyTool)
-        }
-
-        // Model
-        val testModel = LLModel(
-            provider = MockLLMProvider(),
-            id = "test-llm-id",
-            capabilities = emptyList(),
-            contextLength = 1_000,
-        )
-
-        // Prompt
-        val expectedPrompt = Prompt(
-            messages = listOf(
-                systemMessage(systemPrompt),
-                userMessage(userPrompt),
-                assistantMessage(assistantPrompt)
-            ),
-            id = promptId
-        )
-
-        val expectedLLMCallPrompt = expectedPrompt.copy(
-            messages = expectedPrompt.messages
-        )
-
-        // Executor
-        val testLLMResponse = "Default test response"
-
-        val mockExecutor = getMockExecutor {
-            mockLLMAnswer(testLLMResponse).asDefaultResponse onUserRequestEquals userPrompt
-        }
-
-        // Test Data
-        val port = findAvailablePort()
-        val clientConfig = DefaultClientConnectionConfig(host = HOST, port = port, protocol = URLProtocol.HTTP)
-
-        val isClientFinished = CompletableDeferred<Boolean>()
-
-        // Server
-        val serverJob = launch {
-            val strategy = strategy<String, String>("tracing-streaming-success") {
-                val streamAndCollect by nodeLLMRequestStreamingAndSendResults<String>("stream-and-collect")
-
-                edge(nodeStart forwardTo streamAndCollect)
-                edge(
-                    streamAndCollect forwardTo nodeFinish transformed { messages ->
-                        messages.firstOrNull()?.content ?: ""
-                    }
-                )
-            }
-
-            createAgent(
-                agentId = agentId,
-                strategy = strategy,
-                promptId = promptId,
-                userPrompt = userPrompt,
-                systemPrompt = systemPrompt,
-                assistantPrompt = assistantPrompt,
-                toolRegistry = toolRegistry,
-                promptExecutor = mockExecutor,
-                model = testModel,
-            ) {
-                @OptIn(ExperimentalAgentsApi::class)
-                install(Debugger) {
-                    setPort(port)
-
-                    launch {
-                        val messageProcessor = messageProcessors.single() as FeatureMessageRemoteWriter
-                        val isServerStartedCheck = withTimeoutOrNull(defaultClientServerTimeout) {
-                            messageProcessor.isOpen.first { it }
-                        } != null
-
-                        assertTrue(isServerStartedCheck, "Server did not start in time")
-                    }
-                }
-            }.use { agent ->
-                agent.run(userPrompt)
-                isClientFinished.await()
-            }
-        }
-
-        // Client
-        val clientJob = launch {
-            FeatureMessageRemoteClient(
-                connectionConfig = clientConfig,
-                baseClient = testBaseClient,
-                scope = this
-            ).use { client ->
-
-                val clientEventsCollector =
-                    ClientEventsCollector(client = client, expectedEventsCount = 13)
-
-                val collectEventsJob =
-                    clientEventsCollector.startCollectEvents(coroutineScope = this@launch)
-
-                client.connectWithRetry(defaultClientServerTimeout)
-                collectEventsJob.join()
-
-                val callIds = clientEventsCollector.collectedEvents.filterIsInstance<LLMStreamingStartingEvent>().map { it.callId }
-                assertEquals(
-                    1,
-                    callIds.size,
-                    "Expected 1 LLMCallStartingEvent, got ${callIds.size}"
-                )
-
-                // Correct run id will be set after the 'collect events job' is finished.
-                val expectedEvents = listOf(
-                    LLMStreamingStartingEvent(
-                        runId = clientEventsCollector.runId,
-                        callId = callIds[0],
-                        prompt = expectedLLMCallPrompt,
-                        model = testModel.toModelInfo().modelIdentifierName,
-                        tools = listOf(dummyTool.name),
-                        timestamp = testClock.now().toEpochMilliseconds(),
-                    ),
-                    LLMStreamingFrameReceivedEvent(
-                        runId = clientEventsCollector.runId,
-                        callId = callIds[0],
-                        frame = StreamFrame.Append(testLLMResponse),
-                        timestamp = testClock.now().toEpochMilliseconds(),
-                    ),
-                    LLMStreamingCompletedEvent(
-                        runId = clientEventsCollector.runId,
-                        callId = callIds[0],
-                        prompt = expectedLLMCallPrompt,
-                        model = testModel.toModelInfo().modelIdentifierName,
-                        tools = listOf(dummyTool.name),
-                        timestamp = testClock.now().toEpochMilliseconds(),
-                    )
-                )
-
-                val actualEvents = clientEventsCollector.collectedEvents.filter { event ->
-                    event is LLMStreamingStartingEvent ||
-                        event is LLMStreamingFrameReceivedEvent ||
-                        event is LLMStreamingFailedEvent ||
-                        event is LLMStreamingCompletedEvent
-                }
-
-                assertEquals(expectedEvents.size, actualEvents.size)
-                assertContentEquals(expectedEvents, actualEvents)
-
-                isClientFinished.complete(true)
-            }
-        }
-
-        val isFinishedOrNull = withTimeoutOrNull(defaultClientServerTimeout) {
-            listOf(clientJob, serverJob).joinAll()
-        }
-
-        assertNotNull(isFinishedOrNull, "Client or server did not finish in time")
-    }
-
-    @Test
-    fun `test feature message remote writer collect streaming failed events on agent run`() = runBlocking {
-        // Agent Config
-        val agentId = "test-agent-id"
-
-        val userPrompt = "Call the dummy tool with argument: test"
-        val systemPrompt = "Test system prompt"
-        val assistantPrompt = "Test assistant prompt"
-        val promptId = "Test prompt id"
-
-        // Tools
-        val dummyTool = DummyTool()
-
-        val toolRegistry = ToolRegistry {
-            tool(dummyTool)
-        }
-
-        // Model
-        val testModel = LLModel(
-            provider = MockLLMProvider(),
-            id = "test-llm-id",
-            capabilities = emptyList(),
-            contextLength = 1_000,
-        )
-
-        // Prompt
-        val expectedPrompt = Prompt(
-            messages = listOf(
-                systemMessage(systemPrompt),
-                userMessage(userPrompt),
-                assistantMessage(assistantPrompt)
-            ),
-            id = promptId
-        )
-
-        val expectedLLMCallPrompt = expectedPrompt.copy(
-            messages = expectedPrompt.messages
-        )
-
-        // Executor
-        val testStreamingErrorMessage = "Test streaming error"
-        var testStreamingStackTrace = ""
-
-        val testStreamingExecutor = object : PromptExecutor {
-            override suspend fun execute(
-                prompt: Prompt,
-                model: LLModel,
-                tools: List<ToolDescriptor>
-            ): List<Message.Response> = emptyList()
-
-            override fun executeStreaming(
-                prompt: Prompt,
-                model: LLModel,
-                tools: List<ToolDescriptor>
-            ): Flow<StreamFrame> = flow {
-                val testException = IllegalStateException(testStreamingErrorMessage)
-                testStreamingStackTrace = testException.stackTraceToString()
-                throw testException
-            }
-
-            override suspend fun moderate(
-                prompt: Prompt,
-                model: LLModel
-            ): ModerationResult {
-                throw UnsupportedOperationException("Not used in test")
-            }
-
-            override fun close() {}
-        }
-
-        // Test Data
-        val port = findAvailablePort()
-        val clientConfig = DefaultClientConnectionConfig(host = HOST, port = port, protocol = URLProtocol.HTTP)
-
-        val isClientFinished = CompletableDeferred<Boolean>()
-
-        // Server
-        val serverJob = launch {
-            val strategy = strategy<String, String>("tracing-streaming-success") {
-                val streamAndCollect by nodeLLMRequestStreamingAndSendResults<String>("stream-and-collect")
-
-                edge(nodeStart forwardTo streamAndCollect)
-                edge(
-                    streamAndCollect forwardTo nodeFinish transformed { messages ->
-                        messages.firstOrNull()?.content ?: ""
-                    }
-                )
-            }
-
-            createAgent(
-                agentId = agentId,
-                strategy = strategy,
-                promptId = promptId,
-                userPrompt = userPrompt,
-                systemPrompt = systemPrompt,
-                assistantPrompt = assistantPrompt,
-                toolRegistry = toolRegistry,
-                promptExecutor = testStreamingExecutor,
-                model = testModel,
-            ) {
-                @OptIn(ExperimentalAgentsApi::class)
-                install(Debugger) {
-                    setPort(port)
-
-                    launch {
-                        val messageProcessor = messageProcessors.single() as FeatureMessageRemoteWriter
-                        val isServerStartedCheck = withTimeoutOrNull(defaultClientServerTimeout) {
-                            messageProcessor.isOpen.first { it }
-                        } != null
-
-                        assertTrue(isServerStartedCheck, "Server did not start in time")
-                    }
-                }
-            }.use { agent ->
-                val throwable = assertFails {
-                    agent.run(userPrompt)
-                }
-
-                isClientFinished.await()
-
-                assertTrue(throwable is IllegalStateException)
-                assertEquals(testStreamingErrorMessage, throwable.message)
-            }
-        }
-
-        // Client
-        val clientJob = launch {
-            FeatureMessageRemoteClient(
-                connectionConfig = clientConfig,
-                baseClient = testBaseClient,
-                scope = this
-            ).use { client ->
-
-                val clientEventsCollector =
-                    ClientEventsCollector(client = client, expectedEventsCount = 9)
-
-                val collectEventsJob =
-                    clientEventsCollector.startCollectEvents(coroutineScope = this@launch)
-
-                client.connectWithRetry(defaultClientServerTimeout)
-                collectEventsJob.join()
-
-                val callIds = clientEventsCollector.collectedEvents.filterIsInstance<LLMStreamingStartingEvent>().map { it.callId }
-                assertEquals(
-                    1,
-                    callIds.size,
-                    "Expected 1 LLMCallStartingEvent, got ${callIds.size}"
-                )
-
-                // Correct run id will be set after the 'collect events job' is finished.
-                val expectedEvents = listOf(
-                    LLMStreamingStartingEvent(
-                        runId = clientEventsCollector.runId,
-                        callId = callIds[0],
-                        prompt = expectedLLMCallPrompt,
-                        model = testModel.toModelInfo().modelIdentifierName,
-                        tools = listOf(dummyTool.name),
-                        timestamp = testClock.now().toEpochMilliseconds(),
-                    ),
-                    LLMStreamingFailedEvent(
-                        runId = clientEventsCollector.runId,
-                        callId = callIds[0],
-                        error = AIAgentError(testStreamingErrorMessage, testStreamingStackTrace),
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                    LLMStreamingCompletedEvent(
-                        runId = clientEventsCollector.runId,
-                        callId = callIds[0],
-                        prompt = expectedLLMCallPrompt,
-                        model = testModel.toModelInfo().modelIdentifierName,
-                        tools = listOf(dummyTool.name),
-                        timestamp = testClock.now().toEpochMilliseconds(),
-                    )
-                )
-
-                val actualEvents = clientEventsCollector.collectedEvents.filter { event ->
-                    event is LLMStreamingStartingEvent ||
-                        event is LLMStreamingFrameReceivedEvent ||
-                        event is LLMStreamingFailedEvent ||
-                        event is LLMStreamingCompletedEvent
-                }
-
-                assertEquals(expectedEvents.size, actualEvents.size)
-                assertContentEquals(expectedEvents, actualEvents)
-
-                isClientFinished.complete(true)
-            }
-        }
-
-        val isFinishedOrNull = withTimeoutOrNull(defaultClientServerTimeout) {
-            listOf(clientJob, serverJob).joinAll()
-        }
-
-        assertNotNull(isFinishedOrNull, "Client or server did not finish in time")
-    }
-
     //region Port
 
     @Test
@@ -913,7 +499,7 @@ class DebuggerTest {
         @OptIn(ExperimentalAgentsApi::class)
         assertNotNull(portEnvVar, "'${Debugger.KOOG_DEBUGGER_PORT_ENV_VAR}' env variable is not set")
 
-        runAgentPortConfigThroughSystemVariablesTest(portEnvVar.toInt())
+        runAgentPortConfigThroughSystemVariablesTest(port = portEnvVar.toInt())
     }
 
     @Test
@@ -1066,214 +652,4 @@ class DebuggerTest {
     }
 
     //endregion Client Connection Wait Timeout
-
-    //region Private Methods
-
-    private suspend fun runAgentPortConfigThroughSystemVariablesTest(port: Int) = withContext(Dispatchers.Default) {
-        // Agent Config
-        val agentId = "test-agent-id"
-        val strategyName = "test-strategy"
-        val userPrompt = "Call the dummy tool with argument: test"
-
-        val clientConfig = DefaultClientConnectionConfig(
-            host = HOST,
-            port = port,
-            protocol = URLProtocol.HTTP
-        )
-
-        val isClientFinished = CompletableDeferred<Boolean>()
-
-        // Server
-        // The server will read the env variable or VM option to get a port value.
-        val serverJob = launch {
-            val strategy = strategy<String, String>(strategyName) {
-                edge(nodeStart forwardTo nodeFinish)
-            }
-
-            createAgent(
-                agentId = agentId,
-                strategy = strategy,
-                userPrompt = userPrompt,
-            ) {
-                @OptIn(ExperimentalAgentsApi::class)
-                install(Debugger) {
-                    // Do not set the port value explicitly through parameter.
-                    // Use System env var 'KOOG_DEBUGGER_PORT' or VM option 'koog.debugger.port'
-                }
-            }.use { agent ->
-                agent.run(userPrompt)
-                isClientFinished.await()
-            }
-        }
-
-        // Client
-        val clientJob = launch {
-            FeatureMessageRemoteClient(
-                connectionConfig = clientConfig,
-                baseClient = testBaseClient,
-                scope = this
-            ).use { client ->
-
-                val clientEventsCollector = ClientEventsCollector(client = client, expectedEventsCount = 8)
-                val collectEventsJob = clientEventsCollector.startCollectEvents(coroutineScope = this@launch)
-
-                client.connect()
-                collectEventsJob.join()
-
-                val startGraphNode = StrategyEventGraphNode(id = "__start__", name = "__start__")
-                val finishGraphNode = StrategyEventGraphNode(id = "__finish__", name = "__finish__")
-
-                // Correct run id will be set after the 'collect events job' is finished.
-                val expectedEvents = listOf(
-                    AgentStartingEvent(
-                        agentId = agentId,
-                        runId = clientEventsCollector.runId,
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                    GraphStrategyStartingEvent(
-                        runId = clientEventsCollector.runId,
-                        strategyName = strategyName,
-                        graph = StrategyEventGraph(
-                            nodes = listOf(
-                                startGraphNode,
-                                finishGraphNode
-                            ),
-                            edges = listOf(
-                                StrategyEventGraphEdge(startGraphNode, finishGraphNode)
-                            )
-                        ),
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                    NodeExecutionStartingEvent(
-                        runId = clientEventsCollector.runId,
-                        nodeName = "__start__",
-                        input = @OptIn(InternalAgentsApi::class)
-                        SerializationUtils.encodeDataToJsonElementOrNull(
-                            data = userPrompt,
-                            dataType = typeOf<String>()
-                        ),
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                    NodeExecutionCompletedEvent(
-                        runId = clientEventsCollector.runId,
-                        nodeName = "__start__",
-                        input = @OptIn(InternalAgentsApi::class)
-                        SerializationUtils.encodeDataToJsonElementOrNull(
-                            data = userPrompt,
-                            dataType = typeOf<String>()
-                        ),
-                        output = @OptIn(InternalAgentsApi::class)
-                        SerializationUtils.encodeDataToJsonElementOrNull(
-                            data = userPrompt,
-                            dataType = typeOf<String>()
-                        ),
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                    NodeExecutionStartingEvent(
-                        runId = clientEventsCollector.runId,
-                        nodeName = "__finish__",
-                        input = @OptIn(InternalAgentsApi::class)
-                        SerializationUtils.encodeDataToJsonElementOrNull(
-                            data = userPrompt,
-                            dataType = typeOf<String>()
-                        ),
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                    NodeExecutionCompletedEvent(
-                        runId = clientEventsCollector.runId,
-                        nodeName = "__finish__",
-                        input = @OptIn(InternalAgentsApi::class)
-                        SerializationUtils.encodeDataToJsonElementOrNull(
-                            data = userPrompt,
-                            dataType = typeOf<String>()
-                        ),
-                        output = @OptIn(InternalAgentsApi::class)
-                        SerializationUtils.encodeDataToJsonElementOrNull(
-                            data = userPrompt,
-                            dataType = typeOf<String>()
-                        ),
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                    StrategyCompletedEvent(
-                        runId = clientEventsCollector.runId,
-                        strategyName = strategyName,
-                        result = userPrompt,
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                    AgentCompletedEvent(
-                        agentId = agentId,
-                        runId = clientEventsCollector.runId,
-                        result = userPrompt,
-                        timestamp = testClock.now().toEpochMilliseconds()
-                    ),
-                )
-
-                assertEquals(
-                    expectedEvents.size,
-                    clientEventsCollector.collectedEvents.size,
-                    "expectedEventsCount variable in the test need to be updated"
-                )
-                assertContentEquals(expectedEvents, clientEventsCollector.collectedEvents)
-
-                isClientFinished.complete(true)
-            }
-        }
-
-        val isFinishedOrNull = withTimeoutOrNull(defaultClientServerTimeout) {
-            listOf(clientJob, serverJob).joinAll()
-        }
-
-        assertNotNull(isFinishedOrNull, "Client or server did not finish in time")
-    }
-
-    private suspend fun runAgentConnectionWaitConfigThroughSystemVariablesTest(timeout: Duration) = withContext(Dispatchers.Default) {
-        // Agent Config
-        val agentId = "test-agent-id"
-        val strategyName = "test-strategy"
-        val userPrompt = "Call the dummy tool with argument: test"
-
-        // Test Data
-        val port = findAvailablePort()
-        var actualAgentRunTime = Duration.ZERO
-
-        // Server
-        // The server will read the env variable or VM option to get a port value.
-        val serverJob = launch {
-            val strategy = strategy<String, String>(strategyName) {
-                edge(nodeStart forwardTo nodeFinish)
-            }
-
-            createAgent(
-                agentId = agentId,
-                strategy = strategy,
-                userPrompt = userPrompt,
-            ) {
-                @OptIn(ExperimentalAgentsApi::class)
-                install(Debugger) {
-                    setPort(port)
-                    // Do not set the connection awaiting timeout explicitly through parameter.
-                    // Use System env var 'KOOG_DEBUGGER_WAIT_CONNECTION_MS_ENV_VAR' or VM option 'koog.debugger.wait.connection.ms'
-                }
-            }.use { agent ->
-                actualAgentRunTime = measureTime {
-                    withTimeoutOrNull(defaultClientServerTimeout) {
-                        agent.run(userPrompt)
-                    }
-                }
-            }
-        }
-
-        val isFinishedOrNull = withTimeoutOrNull(defaultClientServerTimeout) {
-            serverJob.join()
-        }
-
-        assertNotNull(isFinishedOrNull, "Client or server did not finish in time")
-
-        assertTrue(
-            actualAgentRunTime in timeout..<defaultClientServerTimeout,
-            "Expected actual agent run time is over <$timeout>, but got: <$actualAgentRunTime>"
-        )
-    }
-
-    //endregion Private Methods
 }
